backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
asyncpg>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
aiohttp>=3.11.0
cloudinary>=1.41.0
reportlab>=4.2.2
pillow>=11.1.0
aiofiles>=24.1.0
python-multipart>=0.0.12
pypdf>=5.4.0
